+++ affine/chutes.py
@@ -1,0 +1,278 @@
+
+#!/usr/bin/env python3
+# --------------------------------------------------------------------------- #
+#                             Imports                                         #
+# --------------------------------------------------------------------------- #
+from __future__ import annotations
+import os
+import json
+import time
+import random
+import aiohttp
+import asyncio
+import requests
+from huggingface_hub import HfApi
+from typing import Any, Dict, List, Optional, Union, Tuple, Sequence, Literal, TypeVar, Awaitable
+
+import affine as af
+
+# Model gating check cache
+MODEL_GATING_CACHE = {}  # {model_id: (is_gated, last_checked)}
+_GATING_LOCKS: Dict[int, asyncio.Lock] = {}
+GATING_TTL = 300  # 5 minutes
+
+def _get_gating_lock() -> asyncio.Lock:
+    """Return an asyncio.Lock bound to the current running loop."""
+    try:
+        loop = asyncio.get_running_loop()
+    except RuntimeError:
+        # Fallback if called when no loop is running yet
+        loop = asyncio.get_event_loop()
+    key = id(loop)
+    lock = _GATING_LOCKS.get(key)
+    if lock is None:
+        lock = asyncio.Lock()
+        _GATING_LOCKS[key] = lock
+    return lock
+
+async def check_model_gated(model_id: str, revision: Optional[str] = None) -> Optional[bool]:
+    async with _get_gating_lock():
+        now = time.time()
+        cached = MODEL_GATING_CACHE.get(model_id)
+        if cached and now - cached[1] < GATING_TTL:
+            return cached[0]
+        try:
+            r = await asyncio.to_thread(requests.get, f"https://huggingface.co/api/models/{model_id}", timeout=5)
+            if r.status_code == 200:
+                is_gated = r.json().get("gated", False)
+                if revision:
+                    try:
+                        ok = await asyncio.to_thread(lambda: bool(HfApi(token=os.getenv("HF_TOKEN")).repo_info(repo_id=model_id, revision=revision, repo_type="model")))
+                        if not ok: is_gated = True
+                    except:
+                        pass
+                MODEL_GATING_CACHE[model_id] = (is_gated, now)
+                return is_gated
+        except Exception as e:
+            af.aiohttplogger.trace(f"Gate check failed for {model_id}: {e}")
+        if cached:
+            MODEL_GATING_CACHE[model_id] = (cached[0], now)
+            return cached[0]
+        return None
+
+# --------------------------------------------------------------------------- #
+#                               QUERY                                         #
+# --------------------------------------------------------------------------- #
+# Lazy-initialised semaphore and shared HTTP client
+_HTTP_SEMS: Dict[int, asyncio.Semaphore] = {}
+_CLIENTS: Dict[int, aiohttp.ClientSession] = {}
+
+async def _get_sem() -> asyncio.Semaphore:
+    try:
+        loop = asyncio.get_running_loop()
+    except RuntimeError:
+        loop = asyncio.get_event_loop()
+    key = id(loop)
+    sem = _HTTP_SEMS.get(key)
+    if sem is None:
+        sem = asyncio.Semaphore(int(os.getenv("AFFINE_HTTP_CONCURRENCY", "400")))
+        _HTTP_SEMS[key] = sem
+    return sem
+
+async def _get_client() -> aiohttp.ClientSession:
+    try:
+        loop = asyncio.get_running_loop()
+    except RuntimeError:
+        loop = asyncio.get_event_loop()
+    key = id(loop)
+    client = _CLIENTS.get(key)
+    if client is None or client.closed:
+        limit = int(os.getenv("AFFINE_HTTP_CONCURRENCY", "400"))  # raise this
+        conn = aiohttp.TCPConnector(
+            limit=limit,              # match or exceed your semaphore
+            limit_per_host=0,         # don’t artificially throttle per host
+            ttl_dns_cache=300,        # cache DNS results
+            enable_cleanup_closed=True
+        )
+        client = aiohttp.ClientSession(
+            connector=conn,
+            timeout=aiohttp.ClientTimeout(total=None)
+        )
+        _CLIENTS[key] = client
+    return client
+
+
+TERMINAL = {400, 404, 410}
+async def query(prompt, model: str = "unsloth/gemma-3-12b-it", slug: str = "llm", timeout=150, retries=0, backoff=1) -> af.Response:
+    url = f"https://{slug}.chutes.ai/v1/chat/completions"
+    hdr = {"Authorization": f"Bearer {af.get_conf('CHUTES_API_KEY')}", "Content-Type": "application/json"}
+    start = time.monotonic()
+    af.QCOUNT.labels(model=model).inc()
+    R = lambda resp, at, err, ok: af.Response(response=resp, latency_seconds=time.monotonic()-start,
+                                          attempts=at, model=model, error=err, success=ok)
+    sess = await _get_client()
+    sem = await _get_sem()
+    for attempt in range(1, retries+2):
+        try:
+            payload = {"model": model, "messages": [{"role": "user", "content": prompt}]}
+            async with sem, sess.post(url, json=payload,
+                                      headers=hdr, timeout=timeout) as r:
+                    txt = await r.text(errors="ignore")
+                    if r.status in TERMINAL: return R(None, attempt, f"{r.status}:{txt}", False)
+                    r.raise_for_status()
+                    content = (await r.json())["choices"][0]["message"]["content"]
+                    return R(content, attempt, None, True)
+        except Exception as e:
+            if attempt > retries: return R(None, attempt, str(e), False)
+            await asyncio.sleep(backoff * 2**(attempt-1) * (1 + random.uniform(-0.1, 0.1)))
+
+LOG_TEMPLATE = (
+    "[RESULT] "
+    "{pct:>3.0f}% | "
+    "U{uid:>3d} │ "
+    "{model:<50s} │ "
+    "{env:<3} │ "
+    "{success:^4s} │ "
+    "{score:>6.4f} │ "
+    "{latency:>6.3f}s"
+)
+async def run(challenges, miners, timeout=240, retries=0, backoff=1 )-> List[af.Result]:
+    if not isinstance(challenges, list): challenges = [challenges]
+    if isinstance(miners, af.Miner): miners = [miners]
+    if isinstance(miners, dict):  mmap = miners
+    elif isinstance(miners, list) and all(hasattr(m, "uid") for m in miners):  mmap = {m.uid: m for m in miners}
+    else: mmap = await miners(miners)
+    
+    af.logger.trace("Running challenges: %s on miners: %s", [chal.prompt[:30] for chal in challenges], list(mmap.keys()))
+    response = []
+    
+    async def proc(miner, chal):
+        # Check gating status before querying
+        if miner.model:
+            is_gated = await check_model_gated(miner.model, miner.revision)
+            if is_gated is True:
+                err = "Model is gated"
+                af.logger.trace(f"Miner {miner.uid} - {err} for model {miner.model}")
+                resp = af.Response(response=None, latency_seconds=0, attempts=0, model=miner.model, error=err, success=False)
+                ev = af.Evaluation(env=chal.env, score=0.0, extra={"error": err, "gated": is_gated})
+                return af.Result(miner=miner, challenge=chal, response=resp, evaluation=ev)
+        
+        # Normal processing for non-gated models
+        resp = await query(chal.prompt, miner.model, miner.slug, timeout, retries, backoff)
+        try: ev = await chal.evaluate(resp)
+        except Exception as e: ev = af.Evaluation(env=chal.env, score=0.0, extra={"error": str(e), "evaluation_failed": True})
+        return af.Result(miner=miner, challenge=chal, response=resp, evaluation=ev)
+    
+    tasks = [ asyncio.create_task(proc(m, chal)) for m in mmap.values() if m.model for chal in challenges]  
+    total = len(tasks); completed = 0
+    for task in asyncio.as_completed(tasks): 
+        result: af.Result = await task
+        response.append(result); completed += 1
+        af.logger.debug(
+            LOG_TEMPLATE.format(
+                pct    = completed / total * 100,
+                env    = result.challenge.env.name,                   
+                uid    = result.miner.uid,                 
+                model  = result.miner.model[:50] or "",         
+                success= "RECV" if result.response.success else "NULL",
+                score  = result.evaluation.score,
+                latency= result.response.latency_seconds
+            )
+        )
+    return response
+
+
+# --------------------------------------------------------------------------- #
+#                              Miners                                         #
+# --------------------------------------------------------------------------- #
+async def get_chute(chutes_id: str) -> Dict[str, Any]:
+    url = f"https://api.chutes.ai/chutes/{chutes_id}"
+    token = os.getenv("CHUTES_API_KEY", "")
+    headers = {"Authorization": token}
+    async with aiohttp.ClientSession() as session:
+        async with session.get(url, headers=headers) as r:
+            text = await r.text(errors="ignore")
+            if r.status != 200:
+                return None
+            info = await r.json()
+            for k in ('readme','cords','tagline','instances'):
+                info.pop(k, None)
+            info.get('image', {}).pop('readme', None)
+            return info
+        
+async def get_chute_code(identifier: str) -> Optional[str]:
+    url = f"https://api.chutes.ai/chutes/code/{identifier}"
+    token = os.getenv("CHUTES_API_KEY", "")
+    headers = {"Authorization": token}
+    async with aiohttp.ClientSession() as session:
+        async with session.get(url, headers=headers) as r:
+            if r.status != 200:
+                return None
+            return await r.text(errors="ignore")
+
+async def get_latest_chute_id(model_name: str, api_key: Optional[str] = None) -> Optional[str]:
+    token = api_key or os.getenv("CHUTES_API_KEY", ""); 
+    if not token: return None
+    try:
+        async with aiohttp.ClientSession() as session:
+            async with session.get("https://api.chutes.ai/chutes/", headers={"Authorization": token}) as r:
+                if r.status != 200: return None
+                data = await r.json()
+    except Exception: return None
+    chutes = data.get("items", data) if isinstance(data, dict) else data
+    if not isinstance(chutes, list): return None
+    for chute in reversed(chutes):
+        if any(chute.get(k) == model_name for k in ("model_name", "name", "readme")):
+            return chute.get("chute_id")
+    return None
+
+
+async def get_miners(
+    uids: Optional[Union[int, List[int]]] = None,
+    netuid: int = af.NETUID,
+    meta: object = None,
+) -> Dict[int, af.Miner]:
+    sub = await af.get_subtensor()
+    meta = meta or await sub.metagraph(netuid)
+    commits = await sub.get_all_revealed_commitments(netuid)
+    if uids is None:uids = list(range(len(meta.hotkeys)))
+    elif isinstance(uids, int): uids = [uids]    
+    async def fetch(uid: int):
+        try:
+            hotkey = meta.hotkeys[ uid ]
+            if hotkey not in commits: return None
+            commit = commits[hotkey]
+            block, data = commit[-1]     
+            block = 0 if uid == 0 else block
+            data = json.loads(data)
+            model, miner_revision, chute_id = data.get("model"), data.get("revision"), data.get("chute_id")
+            chute = await get_chute(chute_id)
+            if not chute: None
+            gated = await check_model_gated(model)
+            if gated: return None
+            chutes_name, slug, chutes_revision = chute.get('name'), chute.get("slug"), chute.get("revision")
+            if model != chutes_name or (uid != 0 and chutes_name.split('/')[1].lower()[:6] != 'affine'): return None
+            if chutes_revision == None or miner_revision == chutes_revision:
+                miner = af.Miner(
+                    uid=uid, hotkey=hotkey, model=model, block=int(block),
+                    revision = miner_revision,
+                    slug = slug,
+                    chute=chute,
+                )
+                return miner
+        except: pass
+    results = await asyncio.gather(*(fetch(uid) for uid in uids))
+    output = {uid: m for uid, m in zip(uids, results) if m is not None}
+    # Remove duplicates.
+    if output:
+        best_by_model: Dict[str, Tuple[int, int]] = {}
+        for uid, m in output.items():
+            if not m.model:
+                continue
+            blk = m.block if isinstance(m.block, int) else (int(m.block) if m.block is not None else (2**63 - 1))
+            prev = best_by_model.get(m.model)
+            if prev is None or blk < prev[0]:
+                best_by_model[m.model] = (blk, uid)
+        selected_uids = {uid for _, uid in best_by_model.values()}
+        output = {uid: m for uid, m in output.items() if uid in selected_uids}
+    return output
