+++ affine/dataset.py
@@ -1,0 +1,182 @@
+# --------------------------------------------------------------------------- #
+#                             Imports                                         #
+# --------------------------------------------------------------------------- #
+from __future__ import annotations
+import os
+import json
+import aiohttp
+import asyncio
+import aiofiles
+from .utils import *
+from tqdm import tqdm
+import orjson as _json
+import bittensor as bt
+from pathlib import Path
+from tqdm.asyncio import tqdm
+from botocore.config import Config
+from aiobotocore.session import get_session
+from typing import Any, Dict, List, Optional, Union, Tuple, Sequence, Literal, TypeVar, Awaitable, AsyncIterator
+
+import affine as af
+
+# --------------------------------------------------------------------------- #
+#                             Dataset                                         #
+# --------------------------------------------------------------------------- #
+
+WINDOW        = int(os.getenv("AFFINE_WINDOW", 20))
+RESULT_PREFIX = "affine/results/"
+INDEX_KEY     = "affine/index.json"
+FOLDER  = os.getenv("R2_FOLDER", "affine" )
+BUCKET  = os.getenv("R2_BUCKET_ID", "80f15715bb0b882c9e967c13e677ed7d" )
+ACCESS  = os.getenv("R2_WRITE_ACCESS_KEY_ID", "ff3f4f078019b064bfb6347c270bee4d")
+SECRET  = os.getenv("R2_WRITE_SECRET_ACCESS_KEY", "a94b20516013519b2959cbbb441b9d1ec8511dce3c248223d947be8e85ec754d")
+ENDPOINT = f"https://{BUCKET}.r2.cloudflarestorage.com"
+
+get_client_ctx = lambda: get_session().create_client(
+    "s3", endpoint_url=ENDPOINT,
+    aws_access_key_id=ACCESS, aws_secret_access_key=SECRET,
+    config=Config(max_pool_connections=256)
+)
+CACHE_DIR = Path(os.getenv("AFFINE_CACHE_DIR",
+                 Path.home() / ".cache" / "affine" / "blocks"))
+CACHE_DIR.mkdir(parents=True, exist_ok=True)
+def _w(b: int) -> int: return (b // WINDOW) * WINDOW
+
+# ── fast JSON ───────────────────────────────────────────────────────────────
+_loads, _dumps = _json.loads, _json.dumps
+    
+# ── Index helpers ───────────────────────────────────────────────────────────
+async def _index() -> list[str]:
+    async with get_client_ctx() as c:
+        r = await c.get_object(Bucket=FOLDER, Key=INDEX_KEY)
+        return json.loads(await r["Body"].read())
+
+async def _update_index(k: str) -> None:
+    async with get_client_ctx() as c:
+        try:
+            r = await c.get_object(Bucket=FOLDER, Key=INDEX_KEY)
+            idx = set(json.loads(await r["Body"].read()))
+        except c.exceptions.NoSuchKey:
+            idx = set()
+        if k not in idx:
+            idx.add(k)
+            await c.put_object(Bucket=FOLDER, Key=INDEX_KEY,
+                               Body=_dumps(sorted(idx)),
+                               ContentType="application/json")
+
+# ── Shard cache ─────────────────────────────────────────────────────────────
+async def _cache_shard(key: str, sem: asyncio.Semaphore) -> Path:
+    name, out = Path(key).name, None
+    out = CACHE_DIR / f"{name}.jsonl"; mod = out.with_suffix(".modified")
+    async with sem, get_client_ctx() as c:
+        if out.exists() and mod.exists():
+            h = await c.head_object(Bucket=FOLDER, Key=key)
+            if h["LastModified"].isoformat() == mod.read_text().strip():
+                return out
+        o = await c.get_object(Bucket=FOLDER, Key=key)
+        body, lm = await o["Body"].read(), o["LastModified"].isoformat()
+    tmp = out.with_suffix(".tmp")
+    with tmp.open("wb") as f:
+        f.write(b"\n".join(_dumps(i) for i in _loads(body)) + b"\n")
+    os.replace(tmp, out); mod.write_text(lm)
+    return out
+
+# ── Local JSON‑Lines iterator ───────────────────────────────────────────────
+async def _jsonl(p: Path):
+    async with aiofiles.open(p, "rb") as f:
+        async for l in f: yield l.rstrip(b"\n")
+
+# ── Core async stream (Result objects) ──────────────────────────────────────
+async def rollouts(
+    tail: int,
+    *,
+    max_concurrency: int = 10,      # parallel S3 downloads
+) -> AsyncIterator["af.Result"]:
+    """
+    Stream `Result`s in deterministic order while pre‑downloading future
+    shards concurrently.
+    """
+    # ── figure out which windows we need ────────────────────────────────
+    sub  = await af.get_subtensor()
+    cur  = await sub.get_current_block()
+    need = {w for w in range(_w(cur - tail), _w(cur) + WINDOW, WINDOW)}
+    keys = [
+        k for k in await _index()
+        if (h := Path(k).name.split("-", 1)[0]).isdigit() and int(h) in need
+    ]
+    keys.sort()    
+    # ── helpers ────────────────────────────────
+    sem = asyncio.Semaphore(max_concurrency)     # throttle S3
+    async def _prefetch(key: str) -> Path:       # just downloads / caches
+        return await _cache_shard(key, sem)
+    tasks: list[asyncio.Task[Path]] = [
+        asyncio.create_task(_prefetch(k)) for k in keys[:max_concurrency]
+    ]
+    next_key = max_concurrency            
+    bar = tqdm(f"Dataset=({cur}, {cur - tail})", unit="res", dynamic_ncols=True)
+    # ── main loop: iterate over keys in order ───────────────────────────
+    for i, key in enumerate(keys):
+        path = await tasks[i]
+        if next_key < len(keys):
+            tasks.append(asyncio.create_task(_prefetch(keys[next_key])))
+            next_key += 1
+        async for raw in _jsonl(path):
+            try:
+                r = af.Result.model_validate(_loads(raw))
+                if r.verify():
+                    bar.update(1)
+                    yield r
+            except Exception:
+                pass
+    bar.close()
+    
+    
+# --------------------------------------------------------------------------- #
+async def sign_results( wallet, results ):
+    try:
+        signer_url = af.get_conf('SIGNER_URL', default='http://signer:8080')
+        timeout = aiohttp.ClientTimeout(connect=2, total=30)
+        async with aiohttp.ClientSession(timeout=timeout) as session:
+            payloads = [str(r.challenge) for r in results]
+            resp = await session.post(f"{signer_url}/sign", json={"payloads": payloads})
+            if resp.status == 200:
+                data = await resp.json()
+                sigs = data.get("signatures") or []
+                hotkey = data.get("hotkey")
+                for r, s in zip(results, sigs):
+                    r.hotkey = hotkey
+                    r.signature = s
+    except Exception as e:
+        af.logger.info(f"sink: signer unavailable, using local signing: {type(e).__name__}: {e}")
+        hotkey = wallet.hotkey.ss58_address
+        for r in results: 
+            r.sign(wallet)
+    finally:
+        return hotkey, results
+
+# ── Minimal sink / misc helpers (optional) ──────────────────────────────────
+async def sink(wallet: bt.wallet, results: list["af.Result"], block: int = None):
+    if not results: return
+    if block is None:
+        sub = await af.get_subtensor(); block = await sub.get_current_block()
+    hotkey, signed = await sign_results( wallet, results )
+    key = f"{RESULT_PREFIX}{_w(block):09d}-{hotkey}.json"
+    dumped = [ r.model_dump(mode="json") for r in signed ]
+    async with get_client_ctx() as c:
+        try:
+            r = await c.get_object(Bucket=FOLDER, Key=key)
+            merged = json.loads(await r["Body"].read()) + dumped
+        except c.exceptions.NoSuchKey:
+            merged = dumped
+        await c.put_object(Bucket=FOLDER, Key=key, Body=_dumps(merged),
+                           ContentType="application/json")
+    if len(merged) == len(dumped):              # shard was new
+        await _update_index(key)
+
+async def prune(tail: int):
+    sub = await af.get_subtensor(); cur = await sub.get_current_block()
+    for f in CACHE_DIR.glob("*.jsonl"):
+        b = f.name.split("-", 1)[0]
+        if b.isdigit() and int(b) < cur - tail:
+            try: f.unlink()
+            except OSError: pass
