+++ affine/logging.py
@@ -1,0 +1,65 @@
+
+# --------------------------------------------------------------------------- #
+#                             Imports                                         #
+# --------------------------------------------------------------------------- #
+from __future__ import annotations
+import os
+import logging
+from .utils import *
+from typing import Any, Dict, List, Optional, Union, Tuple, Sequence, Literal, TypeVar, Awaitable
+
+# --------------------------------------------------------------------------- #
+#                       Constants & global singletons                         #
+# --------------------------------------------------------------------------- #
+NETUID = 120
+TRACE  = 5
+logging.addLevelName(TRACE, "TRACE")
+_SINGLETON_CACHE = {}
+def singleton(key:str, factory):
+    """Create a singleton factory function that creates an object only once."""
+    def get_instance():
+        if key not in _SINGLETON_CACHE:
+            _SINGLETON_CACHE[key] = factory()
+        return _SINGLETON_CACHE[key]
+    return get_instance
+
+# --------------------------------------------------------------------------- #
+#                       Prometheus                         #
+# --------------------------------------------------------------------------- #
+from prometheus_client import Counter, CollectorRegistry, start_http_server, Gauge
+METRICS_PORT   = int(os.getenv("AFFINE_METRICS_PORT", "8000"))
+METRICS_ADDR   = os.getenv("AFFINE_METRICS_ADDR", "0.0.0.0")
+REGISTRY       = CollectorRegistry(auto_describe=True)
+QCOUNT  = Counter("qcount", "qcount", ["model"], registry=REGISTRY)
+SCORE   = Gauge( "score", "score", ["uid", "env"], registry=REGISTRY)
+RANK    = Gauge( "rank", "rank", ["uid", "env"], registry=REGISTRY)
+WEIGHT  = Gauge( "weight", "weight", ["uid"], registry=REGISTRY)
+LASTSET = Gauge( "lastset", "lastset", registry=REGISTRY)
+NRESULTS = Gauge( "nresults", "nresults", registry=REGISTRY)
+MAXENV = Gauge("maxenv", "maxenv", ["env"], registry=REGISTRY)
+CACHE = Gauge( "cache", "cache", registry=REGISTRY)
+
+
+# --------------------------------------------------------------------------- #
+#                               Logging                                       #
+# --------------------------------------------------------------------------- #
+def _trace(self, msg, *args, **kwargs):
+    if self.isEnabledFor(TRACE):
+        self._log(TRACE, msg, args, **kwargs)
+logging.Logger.trace = _trace
+logger = logging.getLogger("affine")
+def setup_logging(verbosity: int):
+    if not getattr(setup_logging, "_prom_started", False):
+        try: start_http_server(METRICS_PORT, METRICS_ADDR, registry=REGISTRY)
+        except: pass
+        setup_logging._prom_started = True
+    level = TRACE if verbosity >= 3 else logging.DEBUG if verbosity == 2 else logging.INFO if verbosity == 1 else logging.CRITICAL + 1
+    for noisy in ["websockets", "bittensor", "bittensor-cli", "btdecode", "asyncio", "aiobotocore.regions", "botocore"]:
+        logging.getLogger(noisy).setLevel(logging.WARNING)
+    logging.basicConfig(level=level,
+                        format="%(asctime)s %(levelname)-8s [%(name)s] %(message)s",
+                        datefmt="%Y-%m-%d %H:%M:%S")
+    
+def info():setup_logging(1)
+def debug():setup_logging(2)
+def trace():setup_logging(3)
