+++ pyproject.toml
@@ -23,6 +23,7 @@ dependencies = [
     "botocore>=1.38.27",
     "tabulate>=0.9.0",
     "prometheus-client>=0.21.0",
+    "aiofiles>=24.1.0",
 ]
 readme = "README.md"
 requires-python = ">=3.9,<3.12"
